% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnv_convert.R
\name{.vnv_convert_year}
\alias{.vnv_convert_year}
\title{Helper function to perform CPI adjustment on yearly price data}
\usage{
.vnv_convert_year(obs_value, obs_date, date_unity, include_housing = TRUE)
}
\arguments{
\item{obs_value}{Numeric vector. The price or value you want to convert}

\item{obs_date}{Integer vector, same length as obs_value. The year at which the price/value is recorded}

\item{date_unity}{Integer variable. The year at which you want the CPI to be equal to 1 i.e. the price year. If nothing is input the default value is the most recent CPI release year.}

\item{include_housing}{Boolean variable. Whether to use CPI with housing included or not when adjusting for CPI.}
}
\value{
Returns a numeric vector of length equal to the length of obs_value
}
\description{
Helper function to perform CPI adjustment on yearly price data
}
