% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnv_convert.R
\name{.vnv_date_unity}
\alias{.vnv_date_unity}
\title{Helper function to perform CPI adjustment on monthly price data}
\usage{
.vnv_date_unity(obs_value, obs_date, date_unity, include_housing = TRUE)
}
\arguments{
\item{obs_value}{Numeric vector. The price or value you want to convert}

\item{obs_date}{Date vector, same length as obs_value. The date at which the price/value is recorded}

\item{date_unity}{Date variable. The date at which you want the CPI to be equal to 1 i.e. the price date. If nothing is input the default value is the most recent CPI release date.}

\item{include_housing}{Boolean variable. Whether to use CPI with housing included or not when adjusting for CPI.}
}
\value{
Returns a numeric vector of length equal to the length of obs_value
}
\description{
Helper function to perform CPI adjustment on monthly price data
}
