% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnv.R
\name{vnv}
\alias{vnv}
\title{Fetch consumer price index from Statistics Iceland}
\usage{
vnv(date_unity = NULL, include_housing = TRUE)
}
\arguments{
\item{date_unity}{A date variable of length one at which the CPI takes the value 1. Possible values are dates between 1988-05-01 and the month preceding the current month. The input will be converted to the first of its month to match the CPI release date.}

\item{include_housing}{A boolean to decide whether to include housing in the CPI}
}
\value{
A tibble with two columns denoting the date and the CPI at that date
}
\description{
Fetch consumer price index from Statistics Iceland
}
\examples{
d <- vnv()
d <- vnv(include_housing = FALSE)
my_date <- as.Date("2014-01-15")
d <- vnv(date_unity = my_date)
}
