% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnv_convert.R
\name{vnv_convert}
\alias{vnv_convert}
\title{Convert prices to a common level using icelandic CPI}
\usage{
vnv_convert(obs_value, obs_date, convert_date = NULL, include_housing = TRUE)
}
\arguments{
\item{obs_value}{Numeric vector. The price or value you want to convert}

\item{obs_date}{Date vector, same length as obs_value. The date at which the price/value is recorded}

\item{convert_date}{Date variable. The date at which you want the CPI to be equal to 1 i.e. the price date. If nothing is input the default value is the most recent CPI release date.}

\item{include_housing}{Boolean variable. Whether to use CPI with housing included or not when adjusting for CPI.}
}
\value{
Returns a numeric vector of length equal to the length of obs_value
}
\description{
To allow users to use the most recent CPI as a base, the first time this function is used the CPI is
downloaded from Statistics Iceland and added to the package environment. Then a helper function called
vnv_convert.fun is defined and also added to the package environment.
}
\details{
This means that the CPI timeseries is only downloaded the first time this function is run -hopefully-.
}
\examples{
price <- c(1, 2, 3, 4, 5)
date = as.Date(c("2020-01-01", "2019-01-01", "2018-01-01", "2017-01-01", "2016-01-01"))
price_2020 <- vnv_convert(obs_value = price, obs_date = date, convert_date = as.Date("2020-01-01"))
}
