% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnv.R
\name{vnv_yearly}
\alias{vnv_yearly}
\title{Fetch yearly price index from Statistics Iceland}
\usage{
vnv_yearly(year_unity = NULL, include_housing = TRUE)
}
\arguments{
\item{year_unity}{An integer variable of length one at which the CPI takes the value 1. Possible values are years between 1988 and the current year.}

\item{include_housing}{A boolean to decide whether to include housing in the CPI}
}
\value{
A tibble with two columns denoting the year and the mean CPI at that year
}
\description{
Fetch yearly price index from Statistics Iceland
}
\examples{
d <- vnv_yearly()
d <- vnv_yearly(include_housing = FALSE)
my_year <- 2015
d <- vnv_yearly(year_unity = my_year)
}
